requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
pymupdf==1.27.2.2
pydantic==2.12.5
groq==1.1.2
langgraph==1.1.3
python-dotenv==1.2.2

